requirements.txt
Flask==0.12.2
gunicorn==19.7.1
python-dotenv==0.7.1
coverage==4.4.1
pytest==3.5.1
coveralls==1.3.0